refactor: improve type safety for analysis cache
- Move BenchmarkMetrics and BenchmarkComparison to types/index.ts
- Use proper types instead of unknown in AnalysisCache
- Re-export types from lib/api.ts for backward compatibility
- Remove unnecessary type assertions
